backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
tortoise-orm>=0.21.0,<1.0.0
asyncpg>=0.29.0,<1.0.0
redis>=5.0.0,<6.0.0
pydantic-settings>=2.2.1,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<4.1.0
pyshp>=2.3.1,<3.0.0
